Move encryption service
