+++ src/features/encrypt/EncryptionService.ts
